+++ app/src/main/java/com/hacybeyker/finflow/feature/settings/domain/usecase/SetAppLockEnabledUseCase.kt
@@ -1,0 +1,8 @@
+package com.hacybeyker.finflow.feature.settings.domain.usecase
+
+import com.hacybeyker.finflow.feature.settings.domain.PreferencesRepository
+import javax.inject.Inject
+
+class SetAppLockEnabledUseCase @Inject constructor(private val repository: PreferencesRepository) {
+    suspend operator fun invoke(enabled: Boolean) = repository.setAppLockEnabled(enabled)
+}